64 bit

+++ ConsoleApp2/Program.cs
@@ -9,16 +9,16 @@ namespace ConsoleApp2
         static void Main()
         {
             Int64 num = Convert.ToInt64(Console.ReadLine());
-            var prime_factor = new List<int>();
+            var prime_factor = new List<Int64>();
             for (Int64 i = num; i > 0; i--)
             {
                 decimal remainder = Convert.ToDecimal(num) / Convert.ToDecimal(i);
-                int n = Convert.ToInt32(remainder);
+                Int64 n = Convert.ToInt64(remainder);
                 bool is_a_prime = true;
-                var factor = new List<int>();
+                var factor = new List<Int64>();
                 if (remainder % 1 == 0 && n > 1)
                 {
-                    for (int k = n; k > 0; k--)
+                    for (Int64 k = n; k > 0; k--)
                     { 
                         decimal remain = remainder % Convert.ToDecimal(k);
                         if (remain == 0)
@@ -38,8 +38,8 @@ namespace ConsoleApp2
                     }
                 }
             }
-            int[] primes = prime_factor.ToArray();
-            int largest_prime = primes.Max();
+            Int64[] primes = prime_factor.ToArray();
+            Int64 largest_prime = primes.Max();
             Console.WriteLine($"The largest prime factor is {largest_prime}.");
         }
     }
